# server/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Locadora Objective",
    "description": "API para fazer reservas e locações de filmes",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:3000/api" }],
  "schemes": "http",
  "paths": {
    "/all": {
      "get": {
        "summary": "Retorna uma lista com os filmes disponíveis cadastrados",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Uma lista com os filmes disponíveis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "movies": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "uuid"
                          },
                          "name": {
                            "type": "string"
                          },
                          "synopsis": {
                            "type": "string"
                          },
                          "rating": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/book": {
      "post": {
        "summary": "Cria uma reserva de um filme",
        "produces": ["application/json"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "movieId": {
                    "type": "string",
                    "required": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "um objeto com o id e status da reserva",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reserveId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/confirm": {
      "post": {
        "summary": "Confirmar a locação de um filme",
        "produces": ["application/json"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reserveId": {
                    "type": "string",
                    "required": true
                  },
                  "customer": {
                    "type": "object",
                    "required": true,
                    "properties": {
                      "name": "string",
                      "email": "string",
                      "phone": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "um objeto com o id e status da locação",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scheduleId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/return": {
      "put": {
        "summary": "Devolver um filme",
        "produces": ["application/json"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scheduleId": {
                    "type": "string",
                    "required": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "um objeto com o id e status da locação",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scheduleId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}